Add files via upload
Final submission. This was as complete as I could get it. Sorry it's not quite complete, but I could not find a way to get the starting price nor did I know how to get some of the values or loop it through all the sheets. Can't wait to see the solutions and learn more from here! I know that Google and Stack Overflow have a lot of answers, but I need to get much better at learning how to find them.